Abstract TaskList from main for OOP
